requirements.txt
asgiref==3.2.5
Django==3.0.4
django-appconf==1.0.4
django-cors-headers==3.2.1
django-staticfiles==1.2.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
